feat(packets): make lists and conditional parameters non-optional non-nullable

+++ Packets/NosSmooth.Packets/Client/Inventory/UseItemPacket.cs
@@ -21,5 +21,5 @@ public record UseItemPacket
     [PacketIndex(0)]
     BagType BagType,
     [PacketIndex(1)]
-    short Slot
+    long Slot
 ) : IPacket;
